feat(api): add product telemetry stream endpoint (CHAOS-450)
Ultraworked with [Sisyphus](https://github.com/code-yeongyu/oh-my-openagent)
